requirements.txt
pyyaml>=6.0
requests>=2.28.0
jinja2>=3.1.0

